refactor: create and use AdministrationUserItem
